Observable is now in the Concern module.
