+++ lib/concurrent/concern/obligation.rb
