+++ lib/concurrent/ivar.rb
@@ -2,7 +2,7 @@ require 'thread'
 
 require 'concurrent/errors'
 require 'concurrent/obligation'
-require 'concurrent/observable'
+require 'concurrent/concern/observable'
 require 'concurrent/synchronization'
 
 module Concurrent
@@ -48,7 +48,7 @@ module Concurrent
   #    [DataDrivenFuture in Habanero Java from Rice](http://www.cs.rice.edu/~vs3/hjlib/doc/edu/rice/hj/api/HjDataDrivenFuture.html).
   class IVar < Synchronization::Object
     include Obligation
-    include Observable
+    include Concern::Observable
 
     # @!visibility private
     NO_VALUE = Object.new # :nodoc:
