+++ spec/concurrent/concern/obligation_spec.rb
