+++ spec/concurrent/scheduled_task_spec.rb
@@ -1,7 +1,7 @@
 require 'timecop'
 require_relative 'concern/dereferenceable_shared'
 require_relative 'obligation_shared'
-require_relative 'observable_shared'
+require_relative 'concern/observable_shared'
 
 module Concurrent
